% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-axis.R
\name{Axis}
\alias{Axis}
\title{Axis}
\description{
R6 class representing an axis in a multidimensional execute response.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Axis name (e.g. "Axis0", "SlicerAxis").}

\item{\code{tuples}}{List of tuples. Each tuple is a list of named member info lists.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Axis-new}{\code{Axis$new()}}
\item \href{#method-Axis-print}{\code{Axis$print()}}
\item \href{#method-Axis-clone}{\code{Axis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axis-new"></a>}}
\if{latex}{\out{\hypertarget{method-Axis-new}{}}}
\subsection{Method \code{new()}}{
Create an Axis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axis$new(name = NA_character_, tuples = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Axis name.}

\item{\code{tuples}}{List of tuples.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axis-print"></a>}}
\if{latex}{\out{\hypertarget{method-Axis-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axis$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Axis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Axis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Axis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
