% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap-auth.R
\name{BasicAuth}
\alias{BasicAuth}
\title{BasicAuth authentication}
\description{
BasicAuth authentication

BasicAuth authentication
}
\details{
HTTP Basic authentication.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BasicAuth-new}{\code{BasicAuth$new()}}
\item \href{#method-BasicAuth-apply_auth}{\code{BasicAuth$apply_auth()}}
\item \href{#method-BasicAuth-soap_header_nodes}{\code{BasicAuth$soap_header_nodes()}}
\item \href{#method-BasicAuth-print}{\code{BasicAuth$print()}}
\item \href{#method-BasicAuth-clone}{\code{BasicAuth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicAuth-new"></a>}}
\if{latex}{\out{\hypertarget{method-BasicAuth-new}{}}}
\subsection{Method \code{new()}}{
Create a new BasicAuth object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicAuth$new(username, password)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{The username.}

\item{\code{password}}{The password.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicAuth-apply_auth"></a>}}
\if{latex}{\out{\hypertarget{method-BasicAuth-apply_auth}{}}}
\subsection{Method \code{apply_auth()}}{
Apply Basic Auth header to request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicAuth$apply_auth(req)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{An httr2 request object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The request with Authorization header.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicAuth-soap_header_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-BasicAuth-soap_header_nodes}{}}}
\subsection{Method \code{soap_header_nodes()}}{
Return SOAP header nodes for authentication.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicAuth$soap_header_nodes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL (credentials sent via HTTP only).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicAuth-print"></a>}}
\if{latex}{\out{\hypertarget{method-BasicAuth-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicAuth$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicAuth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BasicAuth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicAuth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
