% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-catalog.R
\name{Catalog}
\alias{Catalog}
\title{Catalog}
\description{
R6 class representing an XMLA catalog.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog_name}}{Name of the catalog.}

\item{\code{description}}{Description.}

\item{\code{roles}}{Roles.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Catalog-new}{\code{Catalog$new()}}
\item \href{#method-Catalog-print}{\code{Catalog$print()}}
\item \href{#method-Catalog-clone}{\code{Catalog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Catalog-new"></a>}}
\if{latex}{\out{\hypertarget{method-Catalog-new}{}}}
\subsection{Method \code{new()}}{
Create a Catalog.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Catalog$new(
  catalog_name = NA_character_,
  description = NA_character_,
  roles = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalog_name}}{Name.}

\item{\code{description}}{Description.}

\item{\code{roles}}{Roles.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Catalog-print"></a>}}
\if{latex}{\out{\hypertarget{method-Catalog-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Catalog$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Catalog-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Catalog-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Catalog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
