% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-catalog.R
\name{Catalog_from_node}
\alias{Catalog_from_node}
\title{Create a Catalog from an XML row node}
\usage{
Catalog_from_node(node)
}
\arguments{
\item{node}{An xml2 node.}
}
\value{
A Catalog object.
}
\description{
Create a Catalog from an XML row node
}
\keyword{internal}
