% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-axis.R
\name{CellData}
\alias{CellData}
\title{CellData}
\description{
R6 class representing cell data in a multidimensional execute response.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cells}}{List of cell lists, each with ordinal, value, formatted_value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CellData-new}{\code{CellData$new()}}
\item \href{#method-CellData-print}{\code{CellData$print()}}
\item \href{#method-CellData-clone}{\code{CellData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CellData-new"></a>}}
\if{latex}{\out{\hypertarget{method-CellData-new}{}}}
\subsection{Method \code{new()}}{
Create CellData.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CellData$new(cells = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cells}}{List of cell lists.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CellData-print"></a>}}
\if{latex}{\out{\hypertarget{method-CellData-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CellData$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CellData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CellData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CellData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
