% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap-connection.R
\name{Connection}
\alias{Connection}
\title{Connection}
\description{
XMLA SOAP connection with optional session support.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{url}}{The XMLA endpoint URL.}

\item{\code{auth}}{The auth object.}

\item{\code{session_id}}{The current session ID, or NULL if stateless.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Connection-new}{\code{Connection$new()}}
\item \href{#method-Connection-start_session}{\code{Connection$start_session()}}
\item \href{#method-Connection-end_session}{\code{Connection$end_session()}}
\item \href{#method-Connection-discover}{\code{Connection$discover()}}
\item \href{#method-Connection-execute}{\code{Connection$execute()}}
\item \href{#method-Connection-print}{\code{Connection$print()}}
\item \href{#method-Connection-clone}{\code{Connection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connection-new"></a>}}
\if{latex}{\out{\hypertarget{method-Connection-new}{}}}
\subsection{Method \code{new()}}{
Create a XMLA connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connection$new(url, auth = NoAuth$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{The endpoint URL.}

\item{\code{auth}}{An auth object (NoAuth or BasicAuth). Defaults to NoAuth.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connection-start_session"></a>}}
\if{latex}{\out{\hypertarget{method-Connection-start_session}{}}}
\subsection{Method \code{start_session()}}{
Start a new XMLA session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connection$start_session()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisible self.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connection-end_session"></a>}}
\if{latex}{\out{\hypertarget{method-Connection-end_session}{}}}
\subsection{Method \code{end_session()}}{
End the current XMLA session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connection$end_session()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisible self.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connection-discover"></a>}}
\if{latex}{\out{\hypertarget{method-Connection-discover}{}}}
\subsection{Method \code{discover()}}{
Send a Discover request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connection$discover(request_type, restrictions = list(), properties = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{request_type}}{The discover request type.}

\item{\code{restrictions}}{Named list of restrictions.}

\item{\code{properties}}{Named list of properties.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An xml2 document of the response.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connection-execute"></a>}}
\if{latex}{\out{\hypertarget{method-Connection-execute}{}}}
\subsection{Method \code{execute()}}{
Send an Execute request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connection$execute(statement, properties = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statement}}{The MDX or DAX or SQL ... statement.}

\item{\code{properties}}{Named list of properties.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An xml2 document of the response.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connection-print"></a>}}
\if{latex}{\out{\hypertarget{method-Connection-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connection$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Connection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
