% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-cube.R
\name{Cube}
\alias{Cube}
\title{Cube}
\description{
R6 class representing an XMLA cube.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog_name}}{Catalog name.}

\item{\code{cube_name}}{Cube name.}

\item{\code{cube_type}}{Cube type.}

\item{\code{description}}{Description.}

\item{\code{last_schema_update}}{Last schema update timestamp.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Cube-new}{\code{Cube$new()}}
\item \href{#method-Cube-print}{\code{Cube$print()}}
\item \href{#method-Cube-clone}{\code{Cube$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cube-new"></a>}}
\if{latex}{\out{\hypertarget{method-Cube-new}{}}}
\subsection{Method \code{new()}}{
Create a Cube.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cube$new(
  catalog_name = NA_character_,
  cube_name = NA_character_,
  cube_type = NA_character_,
  description = NA_character_,
  last_schema_update = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalog_name}}{Catalog.}

\item{\code{cube_name}}{Name.}

\item{\code{cube_type}}{Type.}

\item{\code{description}}{Description.}

\item{\code{last_schema_update}}{Last update.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cube-print"></a>}}
\if{latex}{\out{\hypertarget{method-Cube-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cube$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cube-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Cube-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cube$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
