% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-cube.R
\name{Cube_from_node}
\alias{Cube_from_node}
\title{Create a Cube from an XML row node}
\usage{
Cube_from_node(node)
}
\arguments{
\item{node}{An xml2 node.}
}
\value{
A Cube object.
}
\description{
Create a Cube from an XML row node
}
\keyword{internal}
