% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-datasource.R
\name{DataSource}
\alias{DataSource}
\title{DataSource}
\description{
R6 class representing an XMLA data source.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data_source_name}}{Name of the data source.}

\item{\code{data_source_description}}{Description.}

\item{\code{url}}{URL of the data source.}

\item{\code{provider_name}}{Provider name.}

\item{\code{provider_type}}{Provider type.}

\item{\code{authentication_mode}}{Authentication mode.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSource-new}{\code{DataSource$new()}}
\item \href{#method-DataSource-print}{\code{DataSource$print()}}
\item \href{#method-DataSource-clone}{\code{DataSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataSource-new}{}}}
\subsection{Method \code{new()}}{
Create a DataSource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$new(
  data_source_name = NA_character_,
  data_source_description = NA_character_,
  url = NA_character_,
  provider_name = NA_character_,
  provider_type = NA_character_,
  authentication_mode = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_source_name}}{Name.}

\item{\code{data_source_description}}{Description.}

\item{\code{url}}{URL.}

\item{\code{provider_name}}{Provider.}

\item{\code{provider_type}}{Provider type.}

\item{\code{authentication_mode}}{Auth mode.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSource-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataSource-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
