% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-datasource.R
\name{DataSource_from_node}
\alias{DataSource_from_node}
\title{Create a DataSource from an XML row node}
\usage{
DataSource_from_node(node)
}
\arguments{
\item{node}{An xml2 node.}
}
\value{
A DataSource object.
}
\description{
Create a DataSource from an XML row node
}
\keyword{internal}
