% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-dimension.R
\name{Dimension_from_node}
\alias{Dimension_from_node}
\title{Create a Dimension from an XML row node}
\usage{
Dimension_from_node(node)
}
\arguments{
\item{node}{An xml2 node.}
}
\value{
A Dimension object.
}
\description{
Create a Dimension from an XML row node
}
\keyword{internal}
