% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-discover-helpers.R
\name{discover_catalogs}
\alias{discover_catalogs}
\title{Discover catalogs}
\usage{
discover_catalogs(conn)
}
\arguments{
\item{conn}{A \code{Connection} object.}
}
\value{
A list of \code{Catalog} objects.
}
\description{
Discover catalogs
}
