% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-discover-helpers.R
\name{discover_cubes}
\alias{discover_cubes}
\title{Discover cubes}
\usage{
discover_cubes(conn, catalog)
}
\arguments{
\item{conn}{A \code{Connection} object.}

\item{catalog}{Catalog name.}
}
\value{
A list of \code{Cube} objects.
}
\description{
Discover cubes
}
