% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-discover-helpers.R
\name{discover_datasources}
\alias{discover_datasources}
\title{Discover data sources}
\usage{
discover_datasources(conn)
}
\arguments{
\item{conn}{A \code{Connection} object.}
}
\value{
A list of \code{DataSource} objects.
}
\description{
Discover data sources
}
