% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-discover-helpers.R
\name{discover_dimensions}
\alias{discover_dimensions}
\title{Discover dimensions}
\usage{
discover_dimensions(conn, catalog, cube)
}
\arguments{
\item{conn}{A \code{Connection} object.}

\item{catalog}{Catalog name.}

\item{cube}{Cube name.}
}
\value{
A list of \code{Dimension} objects.
}
\description{
Discover dimensions
}
