% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-discover-helpers.R
\name{discover_hierarchies}
\alias{discover_hierarchies}
\title{Discover hierarchies}
\usage{
discover_hierarchies(conn, catalog, cube, dimension = NULL)
}
\arguments{
\item{conn}{A \code{Connection} object.}

\item{catalog}{Catalog name.}

\item{cube}{Cube name.}

\item{dimension}{Optional dimension unique name filter.}
}
\value{
A list of \code{Hierarchy} objects.
}
\description{
Discover hierarchies
}
