% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-discover-helpers.R
\name{discover_levels}
\alias{discover_levels}
\title{Discover levels}
\usage{
discover_levels(conn, catalog, cube, hierarchy = NULL)
}
\arguments{
\item{conn}{A \code{Connection} object.}

\item{catalog}{Catalog name.}

\item{cube}{Cube name.}

\item{hierarchy}{Optional hierarchy unique name filter.}
}
\value{
A list of \code{Level} objects.
}
\description{
Discover levels
}
