% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-discover-helpers.R
\name{discover_measure_groups}
\alias{discover_measure_groups}
\title{Discover measure groups}
\usage{
discover_measure_groups(conn, catalog, cube)
}
\arguments{
\item{conn}{A \code{Connection} object.}

\item{catalog}{Catalog name.}

\item{cube}{Cube name.}
}
\value{
A data.frame with measure group information.
}
\description{
Discover measure groups
}
