% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-discover-helpers.R
\name{discover_measures}
\alias{discover_measures}
\title{Discover measures}
\usage{
discover_measures(conn, catalog, cube)
}
\arguments{
\item{conn}{A \code{Connection} object.}

\item{catalog}{Catalog name.}

\item{cube}{Cube name.}
}
\value{
A list of \code{Measure} objects.
}
\description{
Discover measures
}
