% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-discover-helpers.R
\name{discover_members}
\alias{discover_members}
\title{Discover members}
\usage{
discover_members(
  conn,
  catalog,
  cube,
  dimension = NULL,
  hierarchy = NULL,
  level = NULL
)
}
\arguments{
\item{conn}{A \code{Connection} object.}

\item{catalog}{Catalog name.}

\item{cube}{Cube name.}

\item{dimension}{Optional dimension unique name filter.}

\item{hierarchy}{Optional hierarchy unique name filter.}

\item{level}{Optional level unique name filter.}
}
\value{
A list of \code{Member} objects.
}
\description{
Discover members
}
