% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-discover-helpers.R
\name{discover_properties}
\alias{discover_properties}
\title{Discover properties}
\usage{
discover_properties(conn, catalog, cube)
}
\arguments{
\item{conn}{A \code{Connection} object.}

\item{catalog}{Catalog name.}

\item{cube}{Cube name.}
}
\value{
A list of \code{Property} objects.
}
\description{
Discover properties
}
