% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-discover-helpers.R
\name{discover_sets}
\alias{discover_sets}
\title{Discover named sets}
\usage{
discover_sets(conn, catalog, cube)
}
\arguments{
\item{conn}{A \code{Connection} object.}

\item{catalog}{Catalog name.}

\item{cube}{Cube name.}
}
\value{
A list of \code{MdSet} objects.
}
\description{
Discover named sets
}
