% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap-objects.R
\name{execute_request}
\alias{execute_request}
\title{Build an XMLA Execute request body node}
\usage{
execute_request(statement, properties = list())
}
\arguments{
\item{statement}{The MDX, DAX, or SQL statement to execute.}

\item{properties}{Named list of property key/value pairs.}
}
\value{
A node descriptor for the Execute element.
}
\description{
Build an XMLA Execute request body node
}
