% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-axis.R
\name{ExecuteResult}
\alias{ExecuteResult}
\title{ExecuteResult}
\description{
R6 class holding the parsed result of a multidimensional Execute response.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{axes}}{List of Axis objects.}

\item{\code{cell_data}}{A CellData object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExecuteResult-new}{\code{ExecuteResult$new()}}
\item \href{#method-ExecuteResult-print}{\code{ExecuteResult$print()}}
\item \href{#method-ExecuteResult-clone}{\code{ExecuteResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecuteResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExecuteResult-new}{}}}
\subsection{Method \code{new()}}{
Create an ExecuteResult.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecuteResult$new(axes = list(), cell_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{axes}}{List of Axis objects.}

\item{\code{cell_data}}{CellData object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecuteResult-print"></a>}}
\if{latex}{\out{\hypertarget{method-ExecuteResult-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecuteResult$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecuteResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExecuteResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecuteResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
