% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-hierarchy.R
\name{Hierarchy}
\alias{Hierarchy}
\title{Hierarchy}
\description{
R6 class representing an XMLA hierarchy.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog_name}}{Catalog name.}

\item{\code{cube_name}}{Cube name.}

\item{\code{dimension_unique_name}}{Dimension unique name.}

\item{\code{hierarchy_name}}{Hierarchy name.}

\item{\code{hierarchy_unique_name}}{Hierarchy unique name.}

\item{\code{hierarchy_caption}}{Hierarchy caption.}

\item{\code{description}}{Description.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Hierarchy-new}{\code{Hierarchy$new()}}
\item \href{#method-Hierarchy-print}{\code{Hierarchy$print()}}
\item \href{#method-Hierarchy-clone}{\code{Hierarchy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hierarchy-new"></a>}}
\if{latex}{\out{\hypertarget{method-Hierarchy-new}{}}}
\subsection{Method \code{new()}}{
Create a Hierarchy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hierarchy$new(
  catalog_name = NA_character_,
  cube_name = NA_character_,
  dimension_unique_name = NA_character_,
  hierarchy_name = NA_character_,
  hierarchy_unique_name = NA_character_,
  hierarchy_caption = NA_character_,
  description = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalog_name}}{Catalog.}

\item{\code{cube_name}}{Cube.}

\item{\code{dimension_unique_name}}{Dimension unique name.}

\item{\code{hierarchy_name}}{Name.}

\item{\code{hierarchy_unique_name}}{Unique name.}

\item{\code{hierarchy_caption}}{Caption.}

\item{\code{description}}{Description.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hierarchy-print"></a>}}
\if{latex}{\out{\hypertarget{method-Hierarchy-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hierarchy$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hierarchy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Hierarchy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hierarchy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
