% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-hierarchy.R
\name{Hierarchy_from_node}
\alias{Hierarchy_from_node}
\title{Create a Hierarchy from an XML row node}
\usage{
Hierarchy_from_node(node)
}
\arguments{
\item{node}{An xml2 node.}
}
\value{
A Hierarchy object.
}
\description{
Create a Hierarchy from an XML row node
}
\keyword{internal}
