% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-level.R
\name{Level}
\alias{Level}
\title{Level}
\description{
R6 class representing an XMLA level.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog_name}}{Catalog name.}

\item{\code{cube_name}}{Cube name.}

\item{\code{dimension_unique_name}}{Dimension unique name.}

\item{\code{hierarchy_unique_name}}{Hierarchy unique name.}

\item{\code{level_name}}{Level name.}

\item{\code{level_unique_name}}{Level unique name.}

\item{\code{level_caption}}{Level caption.}

\item{\code{level_number}}{Level number.}

\item{\code{description}}{Description.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Level-new}{\code{Level$new()}}
\item \href{#method-Level-print}{\code{Level$print()}}
\item \href{#method-Level-clone}{\code{Level$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Level-new"></a>}}
\if{latex}{\out{\hypertarget{method-Level-new}{}}}
\subsection{Method \code{new()}}{
Create a Level.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Level$new(
  catalog_name = NA_character_,
  cube_name = NA_character_,
  dimension_unique_name = NA_character_,
  hierarchy_unique_name = NA_character_,
  level_name = NA_character_,
  level_unique_name = NA_character_,
  level_caption = NA_character_,
  level_number = NA_integer_,
  description = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalog_name}}{Catalog.}

\item{\code{cube_name}}{Cube.}

\item{\code{dimension_unique_name}}{Dimension unique name.}

\item{\code{hierarchy_unique_name}}{Hierarchy unique name.}

\item{\code{level_name}}{Name.}

\item{\code{level_unique_name}}{Unique name.}

\item{\code{level_caption}}{Caption.}

\item{\code{level_number}}{Number.}

\item{\code{description}}{Description.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Level-print"></a>}}
\if{latex}{\out{\hypertarget{method-Level-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Level$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Level-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Level-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Level$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
