% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-level.R
\name{Level_from_node}
\alias{Level_from_node}
\title{Create a Level from an XML row node}
\usage{
Level_from_node(node)
}
\arguments{
\item{node}{An xml2 node.}
}
\value{
A Level object.
}
\description{
Create a Level from an XML row node
}
\keyword{internal}
