% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-set.R
\name{MdSet}
\alias{MdSet}
\title{MdSet}
\description{
R6 class representing an XMLA named set.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog_name}}{Catalog name.}

\item{\code{cube_name}}{Cube name.}

\item{\code{set_name}}{Set name.}

\item{\code{set_caption}}{Set caption.}

\item{\code{scope}}{Scope.}

\item{\code{description}}{Description.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MdSet-new}{\code{MdSet$new()}}
\item \href{#method-MdSet-print}{\code{MdSet$print()}}
\item \href{#method-MdSet-clone}{\code{MdSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MdSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-MdSet-new}{}}}
\subsection{Method \code{new()}}{
Create an MdSet.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MdSet$new(
  catalog_name = NA_character_,
  cube_name = NA_character_,
  set_name = NA_character_,
  set_caption = NA_character_,
  scope = NA_character_,
  description = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalog_name}}{Catalog.}

\item{\code{cube_name}}{Cube.}

\item{\code{set_name}}{Name.}

\item{\code{set_caption}}{Caption.}

\item{\code{scope}}{Scope.}

\item{\code{description}}{Description.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MdSet-print"></a>}}
\if{latex}{\out{\hypertarget{method-MdSet-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MdSet$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MdSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MdSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MdSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
