% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-set.R
\name{MdSet_from_node}
\alias{MdSet_from_node}
\title{Create an MdSet from an XML row node}
\usage{
MdSet_from_node(node)
}
\arguments{
\item{node}{An xml2 node.}
}
\value{
An MdSet object.
}
\description{
Create an MdSet from an XML row node
}
\keyword{internal}
