% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-construction.R
\name{mdx_add_calculated_members}
\alias{mdx_add_calculated_members}
\title{MDX Set Construction Functions (auto-generated)}
\usage{
mdx_add_calculated_members(set_expr)
}
\arguments{
\item{set_expr}{A Set expression.}
}
\value{
MDX string returning Set.
}
\description{
Auto-generated MDX function wrappers for set-construction.
DO NOT EDIT - this file is generated by FunctionMetaDataExportTest.
MDX AddCalculatedMembers function

Adds calculated members to a set.
}
\seealso{
Other set-construction: 
\code{\link{mdx_head}()},
\code{\link{mdx_item}()},
\code{\link{mdx_set_to_str}()},
\code{\link{mdx_strip_calculated_members}()},
\code{\link{mdx_subset}()},
\code{\link{mdx_tail}()}
}
\concept{set-construction}
