% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-aggregate.R
\name{mdx_aggregate}
\alias{mdx_aggregate}
\title{MDX Aggregate Functions (auto-generated)}
\usage{
mdx_aggregate(set_expr, numeric_expr = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{numeric_expr}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Auto-generated MDX function wrappers for aggregate.
DO NOT EDIT - this file is generated by FunctionMetaDataExportTest.
MDX Aggregate function

Returns a calculated value using the appropriate aggregate function, based on the context of the query.
}
\seealso{
Other aggregate: 
\code{\link{mdx_avg}()},
\code{\link{mdx_count}()},
\code{\link{mdx_max}()},
\code{\link{mdx_median}()},
\code{\link{mdx_min}()},
\code{\link{mdx_set_count}()},
\code{\link{mdx_sum}()}
}
\concept{aggregate}
