% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-axis.R
\name{mdx_aggregate_filter}
\alias{mdx_aggregate_filter}
\title{Create an aggregate filter expression for a dimension}
\usage{
mdx_aggregate_filter(dimension, members)
}
\arguments{
\item{dimension}{Dimension unique name.}

\item{members}{Character vector of member unique names.}
}
\value{
MDX Aggregate expression for use in WHERE clause.
}
\description{
Create an aggregate filter expression for a dimension
}
