% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-other.R
\name{mdx_atan2}
\alias{mdx_atan2}
\title{MDX Atan2 function}
\usage{
mdx_atan2(x, y)
}
\arguments{
\item{x}{A Numeric Expression expression.}

\item{y}{A Numeric Expression expression.}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the arctangent, or inverse tangent, of the specified x- and y-coordinates. The arctangent is the angle from the x-axis to a line containing the origin (0, 0) and a point with coordinates (x_num, y_num). The angle is given in radians between -pi and pi, excluding -pi.
}
\seealso{
Other other: 
\code{\link{mdx_acos}()},
\code{\link{mdx_acosh}()},
\code{\link{mdx_asc_b}()},
\code{\link{mdx_asinh}()},
\code{\link{mdx_atanh}()},
\code{\link{mdx_cached_exists}()},
\code{\link{mdx_covariance_n}()},
\code{\link{mdx_current}()},
\code{\link{mdx_current_ordinal}()},
\code{\link{mdx_level_number}()},
\code{\link{mdx_member_caption}()},
\code{\link{mdx_order_key}()},
\code{\link{mdx_param_ref}()},
\code{\link{mdx_parameter}()},
\code{\link{mdx_stddev}()},
\code{\link{mdx_stddev_p}()},
\code{\link{mdx_third_q}()},
\code{\link{mdx_variance}()},
\code{\link{mdx_variance_p}()}
}
\concept{other}
