% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-other.R
\name{mdx_atanh}
\alias{mdx_atanh}
\title{MDX Atanh function}
\usage{
mdx_atanh(number)
}
\arguments{
\item{number}{A Numeric Expression expression.}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the inverse hyperbolic tangent of a number. Number must be between -1 and 1 (excluding -1 and 1).
}
\seealso{
Other other: 
\code{\link{mdx_acos}()},
\code{\link{mdx_acosh}()},
\code{\link{mdx_asc_b}()},
\code{\link{mdx_asinh}()},
\code{\link{mdx_atan2}()},
\code{\link{mdx_cached_exists}()},
\code{\link{mdx_covariance_n}()},
\code{\link{mdx_current}()},
\code{\link{mdx_current_ordinal}()},
\code{\link{mdx_level_number}()},
\code{\link{mdx_member_caption}()},
\code{\link{mdx_order_key}()},
\code{\link{mdx_param_ref}()},
\code{\link{mdx_parameter}()},
\code{\link{mdx_stddev}()},
\code{\link{mdx_stddev_p}()},
\code{\link{mdx_third_q}()},
\code{\link{mdx_variance}()},
\code{\link{mdx_variance_p}()}
}
\concept{other}
