% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-aggregate.R
\name{mdx_avg}
\alias{mdx_avg}
\title{MDX Avg function}
\usage{
mdx_avg(set_expr, numeric_expr = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{numeric_expr}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the average value of a numeric expression evaluated over a set.
}
\seealso{
Other aggregate: 
\code{\link{mdx_aggregate}()},
\code{\link{mdx_count}()},
\code{\link{mdx_max}()},
\code{\link{mdx_median}()},
\code{\link{mdx_min}()},
\code{\link{mdx_set_count}()},
\code{\link{mdx_sum}()}
}
\concept{aggregate}
