% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-member-navigation.R
\name{mdx_calculated_child}
\alias{mdx_calculated_child}
\title{MDX CalculatedChild function}
\usage{
mdx_calculated_child(member, string)
}
\arguments{
\item{member}{A Member expression.}

\item{string}{A String expression.}
}
\value{
MDX string returning Member.
}
\description{
Returns an existing calculated child member with name <String> from the specified <Member>.
}
\seealso{
Other member-navigation: 
\code{\link{mdx_ancestor}()},
\code{\link{mdx_ancestors}()},
\code{\link{mdx_ascendants}()},
\code{\link{mdx_children}()},
\code{\link{mdx_cousin}()},
\code{\link{mdx_data_member}()},
\code{\link{mdx_descendants}()},
\code{\link{mdx_first_child}()},
\code{\link{mdx_first_sibling}()},
\code{\link{mdx_hierarchy_current_member}()},
\code{\link{mdx_hierarchy_default_member}()},
\code{\link{mdx_lag}()},
\code{\link{mdx_last_child}()},
\code{\link{mdx_last_sibling}()},
\code{\link{mdx_lead}()},
\code{\link{mdx_next_member}()},
\code{\link{mdx_parent}()},
\code{\link{mdx_prev_member}()},
\code{\link{mdx_siblings}()},
\code{\link{mdx_str_to_member}()},
\code{\link{mdx_valid_measure}()}
}
\concept{member-navigation}
