% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-helpers.R
\name{mdx_coalesce_empty}
\alias{mdx_coalesce_empty}
\title{MDX CoalesceEmpty function}
\usage{
mdx_coalesce_empty(...)
}
\arguments{
\item{...}{Two or more expressions to coalesce.}
}
\value{
MDX CoalesceEmpty expression string.
}
\description{
Returns the first non-empty expression from the arguments.
}
