% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-statistical.R
\name{mdx_correlation}
\alias{mdx_correlation}
\title{MDX Statistical Functions (auto-generated)}
\usage{
mdx_correlation(set_expr, numeric_expr, numeric_expr_3 = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{numeric_expr}{A Numeric Expression expression.}

\item{numeric_expr_3}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Auto-generated MDX function wrappers for statistical.
DO NOT EDIT - this file is generated by FunctionMetaDataExportTest.
MDX Correlation function

Returns the correlation of two series evaluated over a set.
}
\seealso{
Other statistical: 
\code{\link{mdx_covariance}()},
\code{\link{mdx_lin_reg_intercept}()},
\code{\link{mdx_lin_reg_point}()},
\code{\link{mdx_lin_reg_r2}()},
\code{\link{mdx_lin_reg_slope}()},
\code{\link{mdx_lin_reg_variance}()},
\code{\link{mdx_percentile}()},
\code{\link{mdx_stdev}()},
\code{\link{mdx_stdev_p}()},
\code{\link{mdx_var}()},
\code{\link{mdx_var_p}()}
}
\concept{statistical}
