% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-excel.R
\name{mdx_cosh}
\alias{mdx_cosh}
\title{MDX Excel Functions (auto-generated)}
\usage{
mdx_cosh(number)
}
\arguments{
\item{number}{A Numeric Expression expression.}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Auto-generated MDX function wrappers for excel.
DO NOT EDIT - this file is generated by FunctionMetaDataExportTest.
MDX Cosh function

Returns the hyperbolic cosine of a number.
}
\seealso{
Other excel: 
\code{\link{mdx_degrees}()},
\code{\link{mdx_log10}()},
\code{\link{mdx_mod}()},
\code{\link{mdx_pi}()},
\code{\link{mdx_power}()},
\code{\link{mdx_radians}()},
\code{\link{mdx_sinh}()},
\code{\link{mdx_sqrt_pi}()},
\code{\link{mdx_tanh}()}
}
\concept{excel}
