% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-aggregate.R
\name{mdx_count}
\alias{mdx_count}
\title{MDX Count function}
\usage{
mdx_count(set_expr, pre_post = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{pre_post}{A Symbol expression (optional).}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the number of tuples in a set, empty cells included unless the optional EXCLUDEEMPTY flag is used.
}
\seealso{
Other aggregate: 
\code{\link{mdx_aggregate}()},
\code{\link{mdx_avg}()},
\code{\link{mdx_max}()},
\code{\link{mdx_median}()},
\code{\link{mdx_min}()},
\code{\link{mdx_set_count}()},
\code{\link{mdx_sum}()}
}
\concept{aggregate}
