% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-statistical.R
\name{mdx_covariance}
\alias{mdx_covariance}
\title{MDX Covariance function}
\usage{
mdx_covariance(set_expr, numeric_expression1, numeric_expression2 = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{numeric_expression1}{A Numeric Expression expression.}

\item{numeric_expression2}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the covariance of two series evaluated over a set (biased).
}
\seealso{
Other statistical: 
\code{\link{mdx_correlation}()},
\code{\link{mdx_lin_reg_intercept}()},
\code{\link{mdx_lin_reg_point}()},
\code{\link{mdx_lin_reg_r2}()},
\code{\link{mdx_lin_reg_slope}()},
\code{\link{mdx_lin_reg_variance}()},
\code{\link{mdx_percentile}()},
\code{\link{mdx_stdev}()},
\code{\link{mdx_stdev_p}()},
\code{\link{mdx_var}()},
\code{\link{mdx_var_p}()}
}
\concept{statistical}
