% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-axis.R
\name{mdx_crossjoin}
\alias{mdx_crossjoin}
\title{CrossJoin two or more sets}
\usage{
mdx_crossjoin(...)
}
\arguments{
\item{...}{Set expression strings.}
}
\value{
MDX CrossJoin expression string.
}
\description{
CrossJoin two or more sets
}
