% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-udf.R
\name{mdx_current_date_member}
\alias{mdx_current_date_member}
\title{MDX User-Defined Functions (auto-generated)}
\usage{
mdx_current_date_member(hierarchy, format, match_type = NULL)
}
\arguments{
\item{hierarchy}{A Hierarchy expression.}

\item{format}{A String expression.}

\item{match_type}{A Symbol expression (optional).}
}
\value{
MDX string returning Member.
}
\description{
Auto-generated MDX function wrappers for udf.
DO NOT EDIT - this file is generated by FunctionMetaDataExportTest.
MDX CurrentDateMember function

Returns the closest or exact member within the specified dimension corresponding to the current date, in the format specified by the format parameter. Format strings are the same as used by the MDX Format function, namely the Visual Basic format strings. See http://www.apostate.com/programming/vb-format.html.
}
\seealso{
Other udf: 
\code{\link{mdx_current_date_string}()},
\code{\link{mdx_last_non_empty}()},
\code{\link{mdx_null_value}()}
}
\concept{udf}
