% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-udf.R
\name{mdx_current_date_string}
\alias{mdx_current_date_string}
\title{MDX CurrentDateString function}
\usage{
mdx_current_date_string(format)
}
\arguments{
\item{format}{A String expression.}
}
\value{
MDX string returning String.
}
\description{
Returns the current date formatted as specified by the format parameter.
}
\seealso{
Other udf: 
\code{\link{mdx_current_date_member}()},
\code{\link{mdx_last_non_empty}()},
\code{\link{mdx_null_value}()}
}
\concept{udf}
