% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-excel.R
\name{mdx_degrees}
\alias{mdx_degrees}
\title{MDX Degrees function}
\usage{
mdx_degrees(radians)
}
\arguments{
\item{radians}{A Numeric Expression expression.}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Converts radians to degrees.
}
\seealso{
Other excel: 
\code{\link{mdx_cosh}()},
\code{\link{mdx_log10}()},
\code{\link{mdx_mod}()},
\code{\link{mdx_pi}()},
\code{\link{mdx_power}()},
\code{\link{mdx_radians}()},
\code{\link{mdx_sinh}()},
\code{\link{mdx_sqrt_pi}()},
\code{\link{mdx_tanh}()}
}
\concept{excel}
