% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-member-navigation.R
\name{mdx_descendants}
\alias{mdx_descendants}
\title{MDX Descendants function}
\usage{
mdx_descendants(member, level = NULL, desc_flag = NULL)
}
\arguments{
\item{member}{A Member expression.}

\item{level}{A Level expression (optional).}

\item{desc_flag}{SELF, AFTER, BEFORE, BEFORE_AND_AFTER, SELF_AND_AFTER, SELF_AND_BEFORE, SELF_BEFORE_AFTER, LEAVES.}
}
\value{
MDX string returning Set.
}
\description{
Returns the set of descendants of a member at a specified level, optionally including or excluding descendants in other levels.
}
\seealso{
Other member-navigation: 
\code{\link{mdx_ancestor}()},
\code{\link{mdx_ancestors}()},
\code{\link{mdx_ascendants}()},
\code{\link{mdx_calculated_child}()},
\code{\link{mdx_children}()},
\code{\link{mdx_cousin}()},
\code{\link{mdx_data_member}()},
\code{\link{mdx_first_child}()},
\code{\link{mdx_first_sibling}()},
\code{\link{mdx_hierarchy_current_member}()},
\code{\link{mdx_hierarchy_default_member}()},
\code{\link{mdx_lag}()},
\code{\link{mdx_last_child}()},
\code{\link{mdx_last_sibling}()},
\code{\link{mdx_lead}()},
\code{\link{mdx_next_member}()},
\code{\link{mdx_parent}()},
\code{\link{mdx_prev_member}()},
\code{\link{mdx_siblings}()},
\code{\link{mdx_str_to_member}()},
\code{\link{mdx_valid_measure}()}
}
\concept{member-navigation}
