% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-helpers.R
\name{mdx_diff}
\alias{mdx_diff}
\title{Build expression: difference of two measures}
\usage{
mdx_diff(measure1, measure2)
}
\arguments{
\item{measure1}{First measure unique name.}

\item{measure2}{Second measure unique name.}
}
\value{
MDX expression string.
}
\description{
Build expression: difference of two measures
}
