% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-type-introspection.R,
%   R/mdx-helpers.R
\name{mdx_dimension_name}
\alias{mdx_dimension_name}
\title{MDX Name function}
\usage{
mdx_dimension_name(dimension)

mdx_dimension_name(dimension)
}
\arguments{
\item{dimension}{Dimension expression.}
}
\value{
MDX string returning String.

MDX expression string returning the dimension name.
}
\description{
Returns the name of a dimension.

Returns the name of a dimension.
}
\seealso{
Other type-introspection: 
\code{\link{mdx_dimension_caption}()},
\code{\link{mdx_dimension_unique_name}()},
\code{\link{mdx_dimensions}()},
\code{\link{mdx_hierarchy_all_members}()},
\code{\link{mdx_level_dimension}()},
\code{\link{mdx_level_hierarchy}()},
\code{\link{mdx_level_members}()},
\code{\link{mdx_level_ordinal}()},
\code{\link{mdx_levels}()},
\code{\link{mdx_member_level}()},
\code{\link{mdx_member_unique_name}()},
\code{\link{mdx_members}()},
\code{\link{mdx_value}()}
}
\concept{type-introspection}
