% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-type-introspection.R
\name{mdx_dimensions}
\alias{mdx_dimensions}
\title{MDX Dimensions function}
\usage{
mdx_dimensions(numeric_expression)
}
\arguments{
\item{numeric_expression}{A Numeric Expression expression.}
}
\value{
MDX string returning Hierarchy.
}
\description{
Returns the hierarchy whose zero-based position within the cube is specified by a numeric expression.
}
\seealso{
Other type-introspection: 
\code{\link{mdx_dimension_caption}()},
\code{\link{mdx_dimension_name}()},
\code{\link{mdx_dimension_unique_name}()},
\code{\link{mdx_hierarchy_all_members}()},
\code{\link{mdx_level_dimension}()},
\code{\link{mdx_level_hierarchy}()},
\code{\link{mdx_level_members}()},
\code{\link{mdx_level_ordinal}()},
\code{\link{mdx_levels}()},
\code{\link{mdx_member_level}()},
\code{\link{mdx_member_unique_name}()},
\code{\link{mdx_members}()},
\code{\link{mdx_value}()}
}
\concept{type-introspection}
