% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-operations.R
\name{mdx_distinct}
\alias{mdx_distinct}
\title{MDX Set Operations (auto-generated)}
\usage{
mdx_distinct(set_expr)
}
\arguments{
\item{set_expr}{A Set expression.}
}
\value{
MDX string returning Set.
}
\description{
Auto-generated MDX function wrappers for set-operations.
DO NOT EDIT - this file is generated by FunctionMetaDataExportTest.
MDX Distinct function

Eliminates duplicate tuples from a set.
}
\seealso{
Other set-operations: 
\code{\link{mdx_except}()},
\code{\link{mdx_existing}()},
\code{\link{mdx_exists}()},
\code{\link{mdx_filter}()},
\code{\link{mdx_generate}()},
\code{\link{mdx_hierarchize}()},
\code{\link{mdx_nativize_set}()},
\code{\link{mdx_non_empty}()},
\code{\link{mdx_non_empty_cross_join}()},
\code{\link{mdx_union}()},
\code{\link{mdx_unorder}()}
}
\concept{set-operations}
