% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-drill.R
\name{mdx_drilldown_level}
\alias{mdx_drilldown_level}
\title{MDX Drill Functions (auto-generated)}
\usage{
mdx_drilldown_level(
  set_expr,
  level = NULL,
  index = NULL,
  include_members = NULL
)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{level}{A Level expression (optional).}

\item{index}{A Numeric Expression expression (optional).}

\item{include_members}{A Symbol expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
Auto-generated MDX function wrappers for drill.
DO NOT EDIT - this file is generated by FunctionMetaDataExportTest.
MDX DrilldownLevel function

Drills down the members of a set, at a specified level, to one level below. Alternatively, drills down on a specified dimension in the set.
}
\seealso{
Other drill: 
\code{\link{mdx_drilldown_level_bottom}()},
\code{\link{mdx_drilldown_level_top}()},
\code{\link{mdx_drilldown_member}()},
\code{\link{mdx_toggle_drill_state}()}
}
\concept{drill}
