% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-drill.R
\name{mdx_drilldown_level_top}
\alias{mdx_drilldown_level_top}
\title{MDX DrilldownLevelTop function}
\usage{
mdx_drilldown_level_top(
  set_expr,
  count,
  level = NULL,
  numeric_expression = NULL
)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{count}{A Numeric Expression expression.}

\item{level}{A Level expression (optional).}

\item{numeric_expression}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
Drills down the topmost members of a set, at a specified level, to one level below.
}
\seealso{
Other drill: 
\code{\link{mdx_drilldown_level}()},
\code{\link{mdx_drilldown_level_bottom}()},
\code{\link{mdx_drilldown_member}()},
\code{\link{mdx_toggle_drill_state}()}
}
\concept{drill}
