% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-drill.R
\name{mdx_drilldown_member}
\alias{mdx_drilldown_member}
\title{MDX DrilldownMember function}
\usage{
mdx_drilldown_member(set1, set2, recursive = NULL)
}
\arguments{
\item{set1}{A Set expression.}

\item{set2}{A Set expression.}

\item{recursive}{A Symbol expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
Drills down the members in a set that are present in a second specified set.
}
\seealso{
Other drill: 
\code{\link{mdx_drilldown_level}()},
\code{\link{mdx_drilldown_level_bottom}()},
\code{\link{mdx_drilldown_level_top}()},
\code{\link{mdx_toggle_drill_state}()}
}
\concept{drill}
