% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-operations.R
\name{mdx_filter}
\alias{mdx_filter}
\title{MDX Filter function}
\usage{
mdx_filter(set_expr, logical_expr)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{logical_expr}{A Logical Expression expression.}
}
\value{
MDX string returning Set.
}
\description{
Returns the set resulting from filtering a set based on a search condition.
}
\seealso{
Other set-operations: 
\code{\link{mdx_distinct}()},
\code{\link{mdx_except}()},
\code{\link{mdx_existing}()},
\code{\link{mdx_exists}()},
\code{\link{mdx_generate}()},
\code{\link{mdx_hierarchize}()},
\code{\link{mdx_nativize_set}()},
\code{\link{mdx_non_empty}()},
\code{\link{mdx_non_empty_cross_join}()},
\code{\link{mdx_union}()},
\code{\link{mdx_unorder}()}
}
\concept{set-operations}
