% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-string.R
\name{mdx_format}
\alias{mdx_format}
\title{MDX String Functions (auto-generated)}
\usage{
mdx_format(value, format)
}
\arguments{
\item{value}{A Member expression.}

\item{format}{A String expression.}
}
\value{
MDX string returning String.
}
\description{
Auto-generated MDX function wrappers for string.
DO NOT EDIT - this file is generated by FunctionMetaDataExportTest.
MDX Format function

Formats a number or date to a string.
}
\seealso{
Other string: 
\code{\link{mdx_lcase}()},
\code{\link{mdx_len}()},
\code{\link{mdx_str_to_set}()},
\code{\link{mdx_str_to_tuple}()},
\code{\link{mdx_tuple_to_str}()},
\code{\link{mdx_ucase}()}
}
\concept{string}
