% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-operations.R
\name{mdx_generate}
\alias{mdx_generate}
\title{MDX Generate function}
\usage{
mdx_generate(set1, set2, all = NULL)
}
\arguments{
\item{set1}{A Set expression.}

\item{set2}{A Set expression.}

\item{all}{A Symbol expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
Applies a set to each member of another set and joins the resulting sets by union.
}
\seealso{
Other set-operations: 
\code{\link{mdx_distinct}()},
\code{\link{mdx_except}()},
\code{\link{mdx_existing}()},
\code{\link{mdx_exists}()},
\code{\link{mdx_filter}()},
\code{\link{mdx_hierarchize}()},
\code{\link{mdx_nativize_set}()},
\code{\link{mdx_non_empty}()},
\code{\link{mdx_non_empty_cross_join}()},
\code{\link{mdx_union}()},
\code{\link{mdx_unorder}()}
}
\concept{set-operations}
