% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-operators.R
\name{mdx_greater_equal}
\alias{mdx_greater_equal}
\title{MDX >= function}
\usage{
mdx_greater_equal(numeric1, numeric2)
}
\arguments{
\item{numeric1}{A Numeric Expression expression.}

\item{numeric2}{A Numeric Expression expression.}
}
\value{
MDX string returning Logical Expression.
}
\description{
Returns whether an expression is greater than or equal to another.
}
\seealso{
Other operators: 
\code{\link{mdx_}()},
\code{\link{mdx_and}()},
\code{\link{mdx_divide}()},
\code{\link{mdx_equal}()},
\code{\link{mdx_greater}()},
\code{\link{mdx_in}()},
\code{\link{mdx_is}()},
\code{\link{mdx_is_empty}()},
\code{\link{mdx_is_null}()},
\code{\link{mdx_less}()},
\code{\link{mdx_less_equal}()},
\code{\link{mdx_matches}()},
\code{\link{mdx_minus}()},
\code{\link{mdx_multiply}()},
\code{\link{mdx_not}()},
\code{\link{mdx_not_equal}()},
\code{\link{mdx_or}()},
\code{\link{mdx_plus}()},
\code{\link{mdx_xor}()}
}
\concept{operators}
