% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-construction.R
\name{mdx_head}
\alias{mdx_head}
\title{MDX Head function}
\usage{
mdx_head(set_expr, numeric_expression = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{numeric_expression}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
Returns the first specified number of elements in a set.
}
\seealso{
Other set-construction: 
\code{\link{mdx_add_calculated_members}()},
\code{\link{mdx_item}()},
\code{\link{mdx_set_to_str}()},
\code{\link{mdx_strip_calculated_members}()},
\code{\link{mdx_subset}()},
\code{\link{mdx_tail}()}
}
\concept{set-construction}
