% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-operations.R
\name{mdx_hierarchize}
\alias{mdx_hierarchize}
\title{MDX Hierarchize function}
\usage{
mdx_hierarchize(set_expr, pre_post = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{pre_post}{A Symbol expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
Orders the members of a set in a hierarchy.
}
\seealso{
Other set-operations: 
\code{\link{mdx_distinct}()},
\code{\link{mdx_except}()},
\code{\link{mdx_existing}()},
\code{\link{mdx_exists}()},
\code{\link{mdx_filter}()},
\code{\link{mdx_generate}()},
\code{\link{mdx_nativize_set}()},
\code{\link{mdx_non_empty}()},
\code{\link{mdx_non_empty_cross_join}()},
\code{\link{mdx_union}()},
\code{\link{mdx_unorder}()}
}
\concept{set-operations}
