% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-axis.R
\name{mdx_hierarchy_members}
\alias{mdx_hierarchy_members}
\title{MDX Axis Construction}
\usage{
mdx_hierarchy_members(hierarchy)
}
\arguments{
\item{hierarchy}{Hierarchy unique name string.}
}
\value{
MDX set expression string.
}
\description{
Functions for building MDX axis set expressions.
Get all members of a hierarchy as a set (with braces)
}
