% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-conditional.R
\name{mdx_iif}
\alias{mdx_iif}
\title{MDX Conditional Functions (auto-generated)}
\usage{
mdx_iif(condition, dimension1, dimension2)
}
\arguments{
\item{condition}{A Logical Expression expression.}

\item{dimension1}{A Dimension expression.}

\item{dimension2}{A Dimension expression.}
}
\value{
MDX string returning Dimension.
}
\description{
Auto-generated MDX function wrappers for conditional.
DO NOT EDIT - this file is generated by FunctionMetaDataExportTest.
MDX IIf function

Returns one of two dimension values determined by a logical test.
}
\seealso{
Other conditional: 
\code{\link{mdx_is_empty}()},
\code{\link{mdx_is_null}()}
}
\concept{conditional}
