% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-helpers.R
\name{mdx_intersect}
\alias{mdx_intersect}
\title{MDX Intersect function}
\usage{
mdx_intersect(set1, set2, all = FALSE)
}
\arguments{
\item{set1}{First set expression.}

\item{set2}{Second set expression.}

\item{all}{Optional logical; if TRUE, uses the ALL flag to retain duplicates.}
}
\value{
MDX Intersect expression string.
}
\description{
Returns the intersection of two sets.
}
