% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-conditional.R,
%   R/mdx-functions-operators.R
\name{mdx_is_empty}
\alias{mdx_is_empty}
\title{MDX IsEmpty function}
\usage{
mdx_is_empty(member)

mdx_is_empty(member)
}
\arguments{
\item{member}{A Member expression.}

\item{value}{A String expression.}
}
\value{
MDX string returning Logical Expression.

MDX string returning Logical Expression.
}
\description{
Determines if an expression evaluates to the empty cell value.

A shortcut function for the PeriodsToDate function that specifies the level to be Month.
}
\seealso{
Other conditional: 
\code{\link{mdx_iif}()},
\code{\link{mdx_is_null}()}

Other operators: 
\code{\link{mdx_}()},
\code{\link{mdx_and}()},
\code{\link{mdx_divide}()},
\code{\link{mdx_equal}()},
\code{\link{mdx_greater}()},
\code{\link{mdx_greater_equal}()},
\code{\link{mdx_in}()},
\code{\link{mdx_is}()},
\code{\link{mdx_is_null}()},
\code{\link{mdx_less}()},
\code{\link{mdx_less_equal}()},
\code{\link{mdx_matches}()},
\code{\link{mdx_minus}()},
\code{\link{mdx_multiply}()},
\code{\link{mdx_not}()},
\code{\link{mdx_not_equal}()},
\code{\link{mdx_or}()},
\code{\link{mdx_plus}()},
\code{\link{mdx_xor}()}
}
\concept{conditional}
\concept{operators}
