% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-conditional.R,
%   R/mdx-functions-operators.R
\name{mdx_is_null}
\alias{mdx_is_null}
\title{MDX IsNull function}
\usage{
mdx_is_null(member)

mdx_is_null(member)
}
\arguments{
\item{member}{A Member expression.}

\item{var_name}{A Value expression.}
}
\value{
MDX string returning Logical Expression.

MDX string returning Logical Expression.
}
\description{
Returns a Boolean value that indicates whether an expression contains no valid data (Null).

Returns whether an object is null
}
\seealso{
Other conditional: 
\code{\link{mdx_iif}()},
\code{\link{mdx_is_empty}()}

Other operators: 
\code{\link{mdx_}()},
\code{\link{mdx_and}()},
\code{\link{mdx_divide}()},
\code{\link{mdx_equal}()},
\code{\link{mdx_greater}()},
\code{\link{mdx_greater_equal}()},
\code{\link{mdx_in}()},
\code{\link{mdx_is}()},
\code{\link{mdx_is_empty}()},
\code{\link{mdx_less}()},
\code{\link{mdx_less_equal}()},
\code{\link{mdx_matches}()},
\code{\link{mdx_minus}()},
\code{\link{mdx_multiply}()},
\code{\link{mdx_not}()},
\code{\link{mdx_not_equal}()},
\code{\link{mdx_or}()},
\code{\link{mdx_plus}()},
\code{\link{mdx_xor}()}
}
\concept{conditional}
\concept{operators}
