% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-construction.R
\name{mdx_item}
\alias{mdx_item}
\title{MDX Item function}
\usage{
mdx_item(tuple, index)
}
\arguments{
\item{tuple}{A Tuple expression.}

\item{index}{A Numeric Expression expression.}
}
\value{
MDX string returning Member.
}
\description{
Returns a member from the tuple specified in <Tuple>. The member to be returned is specified by the zero-based position of the member in the set in <Index>.
}
\seealso{
Other set-construction: 
\code{\link{mdx_add_calculated_members}()},
\code{\link{mdx_head}()},
\code{\link{mdx_set_to_str}()},
\code{\link{mdx_strip_calculated_members}()},
\code{\link{mdx_subset}()},
\code{\link{mdx_tail}()}
}
\concept{set-construction}
