% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-udf.R
\name{mdx_last_non_empty}
\alias{mdx_last_non_empty}
\title{MDX LastNonEmpty function}
\usage{
mdx_last_non_empty(set_expr, member)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{member}{A Member expression.}
}
\value{
MDX string returning Member.
}
\description{
Returns the last member of a set whose value is not empty
}
\seealso{
Other udf: 
\code{\link{mdx_current_date_member}()},
\code{\link{mdx_current_date_string}()},
\code{\link{mdx_null_value}()}
}
\concept{udf}
