% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-time.R
\name{mdx_last_periods}
\alias{mdx_last_periods}
\title{MDX LastPeriods function}
\usage{
mdx_last_periods(index, member = NULL)
}
\arguments{
\item{index}{A Numeric Expression expression.}

\item{member}{A Member expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
Returns a set of members prior to and including a specified member.
}
\seealso{
Other time: 
\code{\link{mdx_closing_period}()},
\code{\link{mdx_mtd}()},
\code{\link{mdx_opening_period}()},
\code{\link{mdx_parallel_period}()},
\code{\link{mdx_periods_to_date}()},
\code{\link{mdx_qtd}()},
\code{\link{mdx_wtd}()},
\code{\link{mdx_ytd}()}
}
\concept{time}
