% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-string.R
\name{mdx_lcase}
\alias{mdx_lcase}
\title{MDX LCase function}
\usage{
mdx_lcase(string)
}
\arguments{
\item{string}{A String expression.}
}
\value{
MDX string returning String.
}
\description{
Returns a String that has been converted to lowercase.
}
\seealso{
Other string: 
\code{\link{mdx_format}()},
\code{\link{mdx_len}()},
\code{\link{mdx_str_to_set}()},
\code{\link{mdx_str_to_tuple}()},
\code{\link{mdx_tuple_to_str}()},
\code{\link{mdx_ucase}()}
}
\concept{string}
