% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-string.R
\name{mdx_len}
\alias{mdx_len}
\title{MDX Len function}
\usage{
mdx_len(string)
}
\arguments{
\item{string}{A String expression.}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the number of characters in a string
}
\seealso{
Other string: 
\code{\link{mdx_format}()},
\code{\link{mdx_lcase}()},
\code{\link{mdx_str_to_set}()},
\code{\link{mdx_str_to_tuple}()},
\code{\link{mdx_tuple_to_str}()},
\code{\link{mdx_ucase}()}
}
\concept{string}
