% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-type-introspection.R
\name{mdx_levels}
\alias{mdx_levels}
\title{MDX Levels function}
\usage{
mdx_levels(string)

mdx_levels(string)
}
\arguments{
\item{string}{A String expression.}

\item{hierarchy}{A Hierarchy expression.}
}
\value{
MDX string returning Level.

MDX string returning Level.
}
\description{
Returns the level whose name is specified by a string expression.

Returns the level whose name is specified by a string expression.
}
\seealso{
Other type-introspection: 
\code{\link{mdx_dimension_caption}()},
\code{\link{mdx_dimension_name}()},
\code{\link{mdx_dimension_unique_name}()},
\code{\link{mdx_dimensions}()},
\code{\link{mdx_hierarchy_all_members}()},
\code{\link{mdx_level_dimension}()},
\code{\link{mdx_level_hierarchy}()},
\code{\link{mdx_level_members}()},
\code{\link{mdx_level_ordinal}()},
\code{\link{mdx_member_level}()},
\code{\link{mdx_member_unique_name}()},
\code{\link{mdx_members}()},
\code{\link{mdx_value}()}

Other type-introspection: 
\code{\link{mdx_dimension_caption}()},
\code{\link{mdx_dimension_name}()},
\code{\link{mdx_dimension_unique_name}()},
\code{\link{mdx_dimensions}()},
\code{\link{mdx_hierarchy_all_members}()},
\code{\link{mdx_level_dimension}()},
\code{\link{mdx_level_hierarchy}()},
\code{\link{mdx_level_members}()},
\code{\link{mdx_level_ordinal}()},
\code{\link{mdx_member_level}()},
\code{\link{mdx_member_unique_name}()},
\code{\link{mdx_members}()},
\code{\link{mdx_value}()}
}
\concept{type-introspection}
