% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-statistical.R
\name{mdx_lin_reg_slope}
\alias{mdx_lin_reg_slope}
\title{MDX LinRegSlope function}
\usage{
mdx_lin_reg_slope(set_expr, y, x = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{y}{A Numeric Expression expression.}

\item{x}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Calculates the linear regression of a set and returns the value of a in the regression line y = ax + b.
}
\seealso{
Other statistical: 
\code{\link{mdx_correlation}()},
\code{\link{mdx_covariance}()},
\code{\link{mdx_lin_reg_intercept}()},
\code{\link{mdx_lin_reg_point}()},
\code{\link{mdx_lin_reg_r2}()},
\code{\link{mdx_lin_reg_variance}()},
\code{\link{mdx_percentile}()},
\code{\link{mdx_stdev}()},
\code{\link{mdx_stdev_p}()},
\code{\link{mdx_var}()},
\code{\link{mdx_var_p}()}
}
\concept{statistical}
