% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-aggregate.R
\name{mdx_median}
\alias{mdx_median}
\title{MDX Median function}
\usage{
mdx_median(set_expr, percentile = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{percentile}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the median value of a numeric expression evaluated over a set.
}
\seealso{
Other aggregate: 
\code{\link{mdx_aggregate}()},
\code{\link{mdx_avg}()},
\code{\link{mdx_count}()},
\code{\link{mdx_max}()},
\code{\link{mdx_min}()},
\code{\link{mdx_set_count}()},
\code{\link{mdx_sum}()}
}
\concept{aggregate}
