% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-type-introspection.R
\name{mdx_member_level}
\alias{mdx_member_level}
\title{MDX Level function}
\usage{
mdx_member_level(member)
}
\arguments{
\item{member}{A Member expression.}
}
\value{
MDX string returning Level.
}
\description{
Returns a member's level.
}
\seealso{
Other type-introspection: 
\code{\link{mdx_dimension_caption}()},
\code{\link{mdx_dimension_name}()},
\code{\link{mdx_dimension_unique_name}()},
\code{\link{mdx_dimensions}()},
\code{\link{mdx_hierarchy_all_members}()},
\code{\link{mdx_level_dimension}()},
\code{\link{mdx_level_hierarchy}()},
\code{\link{mdx_level_members}()},
\code{\link{mdx_level_ordinal}()},
\code{\link{mdx_levels}()},
\code{\link{mdx_member_unique_name}()},
\code{\link{mdx_members}()},
\code{\link{mdx_value}()}
}
\concept{type-introspection}
