% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-operators.R
\name{mdx_multiply}
\alias{mdx_multiply}
\title{MDX Operators (auto-generated)}
\usage{
mdx_multiply(numeric1, numeric2)
}
\arguments{
\item{numeric1}{A Set expression.}

\item{numeric2}{A Set expression.}
}
\value{
MDX string returning Set.
}
\description{
Auto-generated MDX function wrappers for operators.
DO NOT EDIT - this file is generated by FunctionMetaDataExportTest.
MDX * function

Returns the cross product of two sets.
}
\seealso{
Other operators: 
\code{\link{mdx_}()},
\code{\link{mdx_and}()},
\code{\link{mdx_divide}()},
\code{\link{mdx_equal}()},
\code{\link{mdx_greater}()},
\code{\link{mdx_greater_equal}()},
\code{\link{mdx_in}()},
\code{\link{mdx_is}()},
\code{\link{mdx_is_empty}()},
\code{\link{mdx_is_null}()},
\code{\link{mdx_less}()},
\code{\link{mdx_less_equal}()},
\code{\link{mdx_matches}()},
\code{\link{mdx_minus}()},
\code{\link{mdx_not}()},
\code{\link{mdx_not_equal}()},
\code{\link{mdx_or}()},
\code{\link{mdx_plus}()},
\code{\link{mdx_xor}()}
}
\concept{operators}
