% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-operations.R
\name{mdx_nativize_set}
\alias{mdx_nativize_set}
\title{MDX NativizeSet function}
\usage{
mdx_nativize_set(set_expr)
}
\arguments{
\item{set_expr}{A Set expression.}
}
\value{
MDX string returning Set.
}
\description{
Tries to natively evaluate <Set>.
}
\seealso{
Other set-operations: 
\code{\link{mdx_distinct}()},
\code{\link{mdx_except}()},
\code{\link{mdx_existing}()},
\code{\link{mdx_exists}()},
\code{\link{mdx_filter}()},
\code{\link{mdx_generate}()},
\code{\link{mdx_hierarchize}()},
\code{\link{mdx_non_empty}()},
\code{\link{mdx_non_empty_cross_join}()},
\code{\link{mdx_union}()},
\code{\link{mdx_unorder}()}
}
\concept{set-operations}
