% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-operations.R
\name{mdx_non_empty}
\alias{mdx_non_empty}
\title{MDX NonEmpty function}
\usage{
mdx_non_empty(set1, set2 = NULL)
}
\arguments{
\item{set1}{A Set expression.}

\item{set2}{A Set expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
Returns the set of tuples that are not empty from a specified set, based on the cross product of the specified set with a second set.
}
\seealso{
Other set-operations: 
\code{\link{mdx_distinct}()},
\code{\link{mdx_except}()},
\code{\link{mdx_existing}()},
\code{\link{mdx_exists}()},
\code{\link{mdx_filter}()},
\code{\link{mdx_generate}()},
\code{\link{mdx_hierarchize}()},
\code{\link{mdx_nativize_set}()},
\code{\link{mdx_non_empty_cross_join}()},
\code{\link{mdx_union}()},
\code{\link{mdx_unorder}()}
}
\concept{set-operations}
