% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-operations.R
\name{mdx_non_empty_cross_join}
\alias{mdx_non_empty_cross_join}
\title{MDX NonEmptyCrossJoin function}
\usage{
mdx_non_empty_cross_join(set1, set2)
}
\arguments{
\item{set1}{A Set expression.}

\item{set2}{A Set expression.}
}
\value{
MDX string returning Set.
}
\description{
Returns the cross product of two sets, excluding empty tuples and tuples without associated fact table data.
}
\seealso{
Other set-operations: 
\code{\link{mdx_distinct}()},
\code{\link{mdx_except}()},
\code{\link{mdx_existing}()},
\code{\link{mdx_exists}()},
\code{\link{mdx_filter}()},
\code{\link{mdx_generate}()},
\code{\link{mdx_hierarchize}()},
\code{\link{mdx_nativize_set}()},
\code{\link{mdx_non_empty}()},
\code{\link{mdx_union}()},
\code{\link{mdx_unorder}()}
}
\concept{set-operations}
