% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-udf.R
\name{mdx_null_value}
\alias{mdx_null_value}
\title{MDX NullValue function}
\usage{
mdx_null_value()
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the null value
}
\seealso{
Other udf: 
\code{\link{mdx_current_date_member}()},
\code{\link{mdx_current_date_string}()},
\code{\link{mdx_last_non_empty}()}
}
\concept{udf}
