% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-time.R
\name{mdx_opening_period}
\alias{mdx_opening_period}
\title{MDX OpeningPeriod function}
\usage{
mdx_opening_period(level = NULL, member = NULL)
}
\arguments{
\item{level}{A Level expression (optional).}

\item{member}{A Member expression (optional).}
}
\value{
MDX string returning Member.
}
\description{
Returns the first descendant of a member at a level.
}
\seealso{
Other time: 
\code{\link{mdx_closing_period}()},
\code{\link{mdx_last_periods}()},
\code{\link{mdx_mtd}()},
\code{\link{mdx_parallel_period}()},
\code{\link{mdx_periods_to_date}()},
\code{\link{mdx_qtd}()},
\code{\link{mdx_wtd}()},
\code{\link{mdx_ytd}()}
}
\concept{time}
