% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-helpers.R
\name{mdx_order}
\alias{mdx_order}
\title{MDX Order function}
\usage{
mdx_order(set, expression, direction = "ASC")
}
\arguments{
\item{set}{Set expression to order.}

\item{expression}{Numeric or string expression to sort by.}

\item{direction}{Sort direction: "ASC", "DESC", "BASC", or "BDESC".}
}
\value{
MDX Order expression string.
}
\description{
Orders members of a set according to an expression.
}
