% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-time.R
\name{mdx_parallel_period}
\alias{mdx_parallel_period}
\title{MDX ParallelPeriod function}
\usage{
mdx_parallel_period(level = NULL, numeric_expr = NULL, member = NULL)
}
\arguments{
\item{level}{A Level expression (optional).}

\item{numeric_expr}{A Numeric Expression expression (optional).}

\item{member}{A Member expression (optional).}
}
\value{
MDX string returning Member.
}
\description{
Returns a member from a prior period in the same relative position as a specified member.
}
\seealso{
Other time: 
\code{\link{mdx_closing_period}()},
\code{\link{mdx_last_periods}()},
\code{\link{mdx_mtd}()},
\code{\link{mdx_opening_period}()},
\code{\link{mdx_periods_to_date}()},
\code{\link{mdx_qtd}()},
\code{\link{mdx_wtd}()},
\code{\link{mdx_ytd}()}
}
\concept{time}
