% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-helpers.R
\name{mdx_pct_of_total}
\alias{mdx_pct_of_total}
\title{Build expression: percentage of total}
\usage{
mdx_pct_of_total(measure, total_member)
}
\arguments{
\item{measure}{Measure unique name.}

\item{total_member}{Member expression for total (e.g., "[Product].[All Products]").}
}
\value{
MDX expression string.
}
\description{
Build expression: percentage of total
}
