% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-time.R
\name{mdx_periods_to_date}
\alias{mdx_periods_to_date}
\title{MDX PeriodsToDate function}
\usage{
mdx_periods_to_date(set_expr = NULL, numeric_expr = NULL)
}
\arguments{
\item{set_expr}{A Set expression (optional).}

\item{numeric_expr}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
Returns a set of periods (members) from a specified level starting with the first period and ending with a specified member.
}
\seealso{
Other time: 
\code{\link{mdx_closing_period}()},
\code{\link{mdx_last_periods}()},
\code{\link{mdx_mtd}()},
\code{\link{mdx_opening_period}()},
\code{\link{mdx_parallel_period}()},
\code{\link{mdx_qtd}()},
\code{\link{mdx_wtd}()},
\code{\link{mdx_ytd}()}
}
\concept{time}
