% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-excel.R
\name{mdx_pi}
\alias{mdx_pi}
\title{MDX Pi function}
\usage{
mdx_pi()
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the number 3.14159265358979, the mathematical constant pi, accurate to 15 digits.
}
\seealso{
Other excel: 
\code{\link{mdx_cosh}()},
\code{\link{mdx_degrees}()},
\code{\link{mdx_log10}()},
\code{\link{mdx_mod}()},
\code{\link{mdx_power}()},
\code{\link{mdx_radians}()},
\code{\link{mdx_sinh}()},
\code{\link{mdx_sqrt_pi}()},
\code{\link{mdx_tanh}()}
}
\concept{excel}
