% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-operators.R
\name{mdx_plus}
\alias{mdx_plus}
\title{MDX + function}
\usage{
mdx_plus(numeric1, numeric1_2)
}
\arguments{
\item{numeric1}{A Numeric Expression expression.}

\item{numeric1_2}{A Numeric Expression expression.}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Adds two numbers.
}
\seealso{
Other operators: 
\code{\link{mdx_}()},
\code{\link{mdx_and}()},
\code{\link{mdx_divide}()},
\code{\link{mdx_equal}()},
\code{\link{mdx_greater}()},
\code{\link{mdx_greater_equal}()},
\code{\link{mdx_in}()},
\code{\link{mdx_is}()},
\code{\link{mdx_is_empty}()},
\code{\link{mdx_is_null}()},
\code{\link{mdx_less}()},
\code{\link{mdx_less_equal}()},
\code{\link{mdx_matches}()},
\code{\link{mdx_minus}()},
\code{\link{mdx_multiply}()},
\code{\link{mdx_not}()},
\code{\link{mdx_not_equal}()},
\code{\link{mdx_or}()},
\code{\link{mdx_xor}()}
}
\concept{operators}
