% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-excel.R
\name{mdx_power}
\alias{mdx_power}
\title{MDX Power function}
\usage{
mdx_power(x, y)
}
\arguments{
\item{x}{A Numeric Expression expression.}

\item{y}{A Numeric Expression expression.}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the result of a number raised to a power.
}
\seealso{
Other excel: 
\code{\link{mdx_cosh}()},
\code{\link{mdx_degrees}()},
\code{\link{mdx_log10}()},
\code{\link{mdx_mod}()},
\code{\link{mdx_pi}()},
\code{\link{mdx_radians}()},
\code{\link{mdx_sinh}()},
\code{\link{mdx_sqrt_pi}()},
\code{\link{mdx_tanh}()}
}
\concept{excel}
