% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-helpers.R
\name{mdx_product}
\alias{mdx_product}
\title{Build expression: product of measures}
\usage{
mdx_product(...)
}
\arguments{
\item{...}{Measure unique names.}
}
\value{
MDX expression string.
}
\description{
Build expression: product of measures
}
