% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-time.R
\name{mdx_qtd}
\alias{mdx_qtd}
\title{MDX Qtd function}
\usage{
mdx_qtd(member = NULL)
}
\arguments{
\item{member}{A Member expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
A shortcut function for the PeriodsToDate function that specifies the level to be Quater.
}
\seealso{
Other time: 
\code{\link{mdx_closing_period}()},
\code{\link{mdx_last_periods}()},
\code{\link{mdx_mtd}()},
\code{\link{mdx_opening_period}()},
\code{\link{mdx_parallel_period}()},
\code{\link{mdx_periods_to_date}()},
\code{\link{mdx_wtd}()},
\code{\link{mdx_ytd}()}
}
\concept{time}
