% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-ranking.R
\name{mdx_rank}
\alias{mdx_rank}
\title{MDX Rank function}
\usage{
mdx_rank(tuple, set_expr, numeric_expr = NULL)
}
\arguments{
\item{tuple}{A Tuple expression.}

\item{set_expr}{A Set expression.}

\item{numeric_expr}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Integer.
}
\description{
Returns the one-based rank of a tuple in a set.
}
\seealso{
Other ranking: 
\code{\link{mdx_bottom_count}()},
\code{\link{mdx_bottom_percent}()},
\code{\link{mdx_bottom_sum}()},
\code{\link{mdx_top_count}()},
\code{\link{mdx_top_percent}()},
\code{\link{mdx_top_sum}()}
}
\concept{ranking}
