% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-helpers.R
\name{mdx_rank_expr}
\alias{mdx_rank_expr}
\title{Build expression: rank within a set}
\usage{
mdx_rank_expr(set, measure)
}
\arguments{
\item{set}{Set expression.}

\item{measure}{Measure to rank by.}
}
\value{
MDX expression string.
}
\description{
Build expression: rank within a set
}
