% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-helpers.R
\name{mdx_ratio}
\alias{mdx_ratio}
\title{Build expression: ratio of two measures}
\usage{
mdx_ratio(numerator, denominator)
}
\arguments{
\item{numerator}{Measure unique name for numerator.}

\item{denominator}{Measure unique name for denominator.}
}
\value{
MDX expression string.
}
\description{
Build expression: ratio of two measures
}
