% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-helpers.R
\name{mdx_running_sum}
\alias{mdx_running_sum}
\title{Build expression: running sum / cumulative total}
\usage{
mdx_running_sum(measure, hierarchy)
}
\arguments{
\item{measure}{Measure unique name.}

\item{hierarchy}{Hierarchy unique name for the running dimension.}
}
\value{
MDX expression string.
}
\description{
Build expression: running sum / cumulative total
}
