% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-aggregate.R
\name{mdx_set_count}
\alias{mdx_set_count}
\title{MDX Count function}
\usage{
mdx_set_count(set_expr)
}
\arguments{
\item{set_expr}{A Set expression.}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the number of tuples in a set including empty cells.
}
\seealso{
Other aggregate: 
\code{\link{mdx_aggregate}()},
\code{\link{mdx_avg}()},
\code{\link{mdx_count}()},
\code{\link{mdx_max}()},
\code{\link{mdx_median}()},
\code{\link{mdx_min}()},
\code{\link{mdx_sum}()}
}
\concept{aggregate}
