% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-construction.R
\name{mdx_set_to_str}
\alias{mdx_set_to_str}
\title{MDX SetToStr function}
\usage{
mdx_set_to_str(set_expr)
}
\arguments{
\item{set_expr}{A Set expression.}
}
\value{
MDX string returning String.
}
\description{
Constructs a string from a set.
}
\seealso{
Other set-construction: 
\code{\link{mdx_add_calculated_members}()},
\code{\link{mdx_head}()},
\code{\link{mdx_item}()},
\code{\link{mdx_strip_calculated_members}()},
\code{\link{mdx_subset}()},
\code{\link{mdx_tail}()}
}
\concept{set-construction}
