% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-excel.R
\name{mdx_sinh}
\alias{mdx_sinh}
\title{MDX Sinh function}
\usage{
mdx_sinh(number)
}
\arguments{
\item{number}{A Numeric Expression expression.}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the hyperbolic sine of a number.
}
\seealso{
Other excel: 
\code{\link{mdx_cosh}()},
\code{\link{mdx_degrees}()},
\code{\link{mdx_log10}()},
\code{\link{mdx_mod}()},
\code{\link{mdx_pi}()},
\code{\link{mdx_power}()},
\code{\link{mdx_radians}()},
\code{\link{mdx_sqrt_pi}()},
\code{\link{mdx_tanh}()}
}
\concept{excel}
