% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-vba.R
\name{mdx_sqr}
\alias{mdx_sqr}
\title{MDX Sqr function}
\usage{
mdx_sqr(number)
}
\arguments{
\item{number}{A Numeric Expression expression.}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns a Double specifying the square root of a number.
}
\seealso{
Other vba: 
\code{\link{mdx_abs}()},
\code{\link{mdx_asc}()},
\code{\link{mdx_atn}()},
\code{\link{mdx_cbool}()},
\code{\link{mdx_cbyte}()},
\code{\link{mdx_cdate}()},
\code{\link{mdx_cdbl}()},
\code{\link{mdx_chr}()},
\code{\link{mdx_chr_b}()},
\code{\link{mdx_cint}()},
\code{\link{mdx_cos}()},
\code{\link{mdx_date}()},
\code{\link{mdx_date_add}()},
\code{\link{mdx_date_diff}()},
\code{\link{mdx_date_part}()},
\code{\link{mdx_date_serial}()},
\code{\link{mdx_date_value}()},
\code{\link{mdx_day}()},
\code{\link{mdx_ddb}()},
\code{\link{mdx_exp}()},
\code{\link{mdx_fix}()},
\code{\link{mdx_format_currency}()},
\code{\link{mdx_format_date_time}()},
\code{\link{mdx_format_number}()},
\code{\link{mdx_format_percent}()},
\code{\link{mdx_fv}()},
\code{\link{mdx_hex}()},
\code{\link{mdx_hour}()},
\code{\link{mdx_in_str}()},
\code{\link{mdx_in_str_rev}()},
\code{\link{mdx_int}()},
\code{\link{mdx_ipmt}()},
\code{\link{mdx_irr}()},
\code{\link{mdx_is_array}()},
\code{\link{mdx_is_date}()},
\code{\link{mdx_is_error}()},
\code{\link{mdx_is_missing}()},
\code{\link{mdx_is_numeric}()},
\code{\link{mdx_is_object}()},
\code{\link{mdx_left}()},
\code{\link{mdx_log}()},
\code{\link{mdx_ltrim}()},
\code{\link{mdx_mid}()},
\code{\link{mdx_minute}()},
\code{\link{mdx_mirr}()},
\code{\link{mdx_month}()},
\code{\link{mdx_month_name}()},
\code{\link{mdx_now}()},
\code{\link{mdx_nper}()},
\code{\link{mdx_npv}()},
\code{\link{mdx_oct}()},
\code{\link{mdx_pmt}()},
\code{\link{mdx_ppmt}()},
\code{\link{mdx_pv}()},
\code{\link{mdx_rate}()},
\code{\link{mdx_replace}()},
\code{\link{mdx_right}()},
\code{\link{mdx_round}()},
\code{\link{mdx_rtrim}()},
\code{\link{mdx_second}()},
\code{\link{mdx_sgn}()},
\code{\link{mdx_sin}()},
\code{\link{mdx_sln}()},
\code{\link{mdx_space}()},
\code{\link{mdx_str}()},
\code{\link{mdx_str_comp}()},
\code{\link{mdx_str_reverse}()},
\code{\link{mdx_string}()},
\code{\link{mdx_syd}()},
\code{\link{mdx_tan}()},
\code{\link{mdx_time}()},
\code{\link{mdx_time_serial}()},
\code{\link{mdx_time_value}()},
\code{\link{mdx_timer}()},
\code{\link{mdx_trim}()},
\code{\link{mdx_type_name}()},
\code{\link{mdx_val}()},
\code{\link{mdx_weekday}()},
\code{\link{mdx_weekday_name}()},
\code{\link{mdx_year}()}
}
\concept{vba}
