% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-other.R
\name{mdx_stddev}
\alias{mdx_stddev}
\title{MDX Stddev function}
\usage{
mdx_stddev(set_expr, numeric_expr = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{numeric_expr}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Alias for Stdev.
}
\seealso{
Other other: 
\code{\link{mdx_acos}()},
\code{\link{mdx_acosh}()},
\code{\link{mdx_asc_b}()},
\code{\link{mdx_asinh}()},
\code{\link{mdx_atan2}()},
\code{\link{mdx_atanh}()},
\code{\link{mdx_cached_exists}()},
\code{\link{mdx_covariance_n}()},
\code{\link{mdx_current}()},
\code{\link{mdx_current_ordinal}()},
\code{\link{mdx_level_number}()},
\code{\link{mdx_member_caption}()},
\code{\link{mdx_order_key}()},
\code{\link{mdx_param_ref}()},
\code{\link{mdx_parameter}()},
\code{\link{mdx_stddev_p}()},
\code{\link{mdx_third_q}()},
\code{\link{mdx_variance}()},
\code{\link{mdx_variance_p}()}
}
\concept{other}
