% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-statistical.R
\name{mdx_stdev}
\alias{mdx_stdev}
\title{MDX Stdev function}
\usage{
mdx_stdev(set_expr, numeric_expr = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{numeric_expr}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the standard deviation of a numeric expression evaluated over a set (unbiased).
}
\seealso{
Other statistical: 
\code{\link{mdx_correlation}()},
\code{\link{mdx_covariance}()},
\code{\link{mdx_lin_reg_intercept}()},
\code{\link{mdx_lin_reg_point}()},
\code{\link{mdx_lin_reg_r2}()},
\code{\link{mdx_lin_reg_slope}()},
\code{\link{mdx_lin_reg_variance}()},
\code{\link{mdx_percentile}()},
\code{\link{mdx_stdev_p}()},
\code{\link{mdx_var}()},
\code{\link{mdx_var_p}()}
}
\concept{statistical}
