% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-string.R
\name{mdx_str_to_tuple}
\alias{mdx_str_to_tuple}
\title{MDX StrToTuple function}
\usage{
mdx_str_to_tuple(string_expr)
}
\arguments{
\item{string_expr}{A String expression.}
}
\value{
MDX string returning Tuple.
}
\description{
Constructs a tuple from a string.
}
\seealso{
Other string: 
\code{\link{mdx_format}()},
\code{\link{mdx_lcase}()},
\code{\link{mdx_len}()},
\code{\link{mdx_str_to_set}()},
\code{\link{mdx_tuple_to_str}()},
\code{\link{mdx_ucase}()}
}
\concept{string}
