% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-construction.R
\name{mdx_strip_calculated_members}
\alias{mdx_strip_calculated_members}
\title{MDX StripCalculatedMembers function}
\usage{
mdx_strip_calculated_members(set_expr)
}
\arguments{
\item{set_expr}{A Set expression.}
}
\value{
MDX string returning Set.
}
\description{
Removes calculated members from a set.
}
\seealso{
Other set-construction: 
\code{\link{mdx_add_calculated_members}()},
\code{\link{mdx_head}()},
\code{\link{mdx_item}()},
\code{\link{mdx_set_to_str}()},
\code{\link{mdx_subset}()},
\code{\link{mdx_tail}()}
}
\concept{set-construction}
