% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-construction.R
\name{mdx_subset}
\alias{mdx_subset}
\title{MDX Subset function}
\usage{
mdx_subset(set_expr, start, count = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{start}{A Numeric Expression expression.}

\item{count}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
Returns a subset of elements from a set.
}
\seealso{
Other set-construction: 
\code{\link{mdx_add_calculated_members}()},
\code{\link{mdx_head}()},
\code{\link{mdx_item}()},
\code{\link{mdx_set_to_str}()},
\code{\link{mdx_strip_calculated_members}()},
\code{\link{mdx_tail}()}
}
\concept{set-construction}
