% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-helpers.R
\name{mdx_sum_expr}
\alias{mdx_sum_expr}
\title{Build expression: sum of measures}
\usage{
mdx_sum_expr(...)
}
\arguments{
\item{...}{Measure unique names.}
}
\value{
MDX expression string.
}
\description{
Build expression: sum of measures
}
