% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-construction.R
\name{mdx_tail}
\alias{mdx_tail}
\title{MDX Tail function}
\usage{
mdx_tail(set_expr, count = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{count}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
Returns a subset from the end of a set.
}
\seealso{
Other set-construction: 
\code{\link{mdx_add_calculated_members}()},
\code{\link{mdx_head}()},
\code{\link{mdx_item}()},
\code{\link{mdx_set_to_str}()},
\code{\link{mdx_strip_calculated_members}()},
\code{\link{mdx_subset}()}
}
\concept{set-construction}
