% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-excel.R
\name{mdx_tanh}
\alias{mdx_tanh}
\title{MDX Tanh function}
\usage{
mdx_tanh(radians)
}
\arguments{
\item{radians}{A Numeric Expression expression.}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the hyperbolic tangent of a number.
}
\seealso{
Other excel: 
\code{\link{mdx_cosh}()},
\code{\link{mdx_degrees}()},
\code{\link{mdx_log10}()},
\code{\link{mdx_mod}()},
\code{\link{mdx_pi}()},
\code{\link{mdx_power}()},
\code{\link{mdx_radians}()},
\code{\link{mdx_sinh}()},
\code{\link{mdx_sqrt_pi}()}
}
\concept{excel}
