% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-other.R
\name{mdx_third_q}
\alias{mdx_third_q}
\title{MDX ThirdQ function}
\usage{
mdx_third_q(set_expr, range = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{range}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Numeric Expression.
}
\description{
Returns the 3rd quartile value of a numeric expression evaluated over a set.
}
\seealso{
Other other: 
\code{\link{mdx_acos}()},
\code{\link{mdx_acosh}()},
\code{\link{mdx_asc_b}()},
\code{\link{mdx_asinh}()},
\code{\link{mdx_atan2}()},
\code{\link{mdx_atanh}()},
\code{\link{mdx_cached_exists}()},
\code{\link{mdx_covariance_n}()},
\code{\link{mdx_current}()},
\code{\link{mdx_current_ordinal}()},
\code{\link{mdx_level_number}()},
\code{\link{mdx_member_caption}()},
\code{\link{mdx_order_key}()},
\code{\link{mdx_param_ref}()},
\code{\link{mdx_parameter}()},
\code{\link{mdx_stddev}()},
\code{\link{mdx_stddev_p}()},
\code{\link{mdx_variance}()},
\code{\link{mdx_variance_p}()}
}
\concept{other}
