% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-drill.R
\name{mdx_toggle_drill_state}
\alias{mdx_toggle_drill_state}
\title{MDX ToggleDrillState function}
\usage{
mdx_toggle_drill_state(set1, set2, recursive = NULL)
}
\arguments{
\item{set1}{A Set expression.}

\item{set2}{A Set expression.}

\item{recursive}{A Symbol expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
Toggles the drill state of members. This function is a combination of DrillupMember and DrilldownMember.
}
\seealso{
Other drill: 
\code{\link{mdx_drilldown_level}()},
\code{\link{mdx_drilldown_level_bottom}()},
\code{\link{mdx_drilldown_level_top}()},
\code{\link{mdx_drilldown_member}()}
}
\concept{drill}
