% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-ranking.R
\name{mdx_top_count}
\alias{mdx_top_count}
\title{MDX TopCount function}
\usage{
mdx_top_count(set_expr, count, numeric_expr = NULL)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{count}{A Numeric Expression expression.}

\item{numeric_expr}{A Numeric Expression expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
Returns a specified number of items from the top of a set, optionally ordering the set first.
}
\seealso{
Other ranking: 
\code{\link{mdx_bottom_count}()},
\code{\link{mdx_bottom_percent}()},
\code{\link{mdx_bottom_sum}()},
\code{\link{mdx_rank}()},
\code{\link{mdx_top_percent}()},
\code{\link{mdx_top_sum}()}
}
\concept{ranking}
