% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-ranking.R
\name{mdx_top_sum}
\alias{mdx_top_sum}
\title{MDX TopSum function}
\usage{
mdx_top_sum(set_expr, value, numeric_expr)
}
\arguments{
\item{set_expr}{A Set expression.}

\item{value}{A Numeric Expression expression.}

\item{numeric_expr}{A Numeric Expression expression.}
}
\value{
MDX string returning Set.
}
\description{
Sorts a set and returns the top N elements whose cumulative total is at least a specified value.
}
\seealso{
Other ranking: 
\code{\link{mdx_bottom_count}()},
\code{\link{mdx_bottom_percent}()},
\code{\link{mdx_bottom_sum}()},
\code{\link{mdx_rank}()},
\code{\link{mdx_top_count}()},
\code{\link{mdx_top_percent}()}
}
\concept{ranking}
