% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-string.R
\name{mdx_tuple_to_str}
\alias{mdx_tuple_to_str}
\title{MDX TupleToStr function}
\usage{
mdx_tuple_to_str(tuple)
}
\arguments{
\item{tuple}{A Tuple expression.}
}
\value{
MDX string returning String.
}
\description{
Constructs a string from a tuple.
}
\seealso{
Other string: 
\code{\link{mdx_format}()},
\code{\link{mdx_lcase}()},
\code{\link{mdx_len}()},
\code{\link{mdx_str_to_set}()},
\code{\link{mdx_str_to_tuple}()},
\code{\link{mdx_ucase}()}
}
\concept{string}
