% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-string.R
\name{mdx_ucase}
\alias{mdx_ucase}
\title{MDX UCase function}
\usage{
mdx_ucase(string)
}
\arguments{
\item{string}{A String expression.}
}
\value{
MDX string returning String.
}
\description{
Returns a string that has been converted to uppercase
}
\seealso{
Other string: 
\code{\link{mdx_format}()},
\code{\link{mdx_lcase}()},
\code{\link{mdx_len}()},
\code{\link{mdx_str_to_set}()},
\code{\link{mdx_str_to_tuple}()},
\code{\link{mdx_tuple_to_str}()}
}
\concept{string}
