% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-functions-set-operations.R
\name{mdx_union}
\alias{mdx_union}
\title{MDX Union function}
\usage{
mdx_union(set1, set2, all = NULL)
}
\arguments{
\item{set1}{A Set expression.}

\item{set2}{A Set expression.}

\item{all}{A Symbol expression (optional).}
}
\value{
MDX string returning Set.
}
\description{
Returns the union of two sets, optionally retaining duplicates.
}
\seealso{
Other set-operations: 
\code{\link{mdx_distinct}()},
\code{\link{mdx_except}()},
\code{\link{mdx_existing}()},
\code{\link{mdx_exists}()},
\code{\link{mdx_filter}()},
\code{\link{mdx_generate}()},
\code{\link{mdx_hierarchize}()},
\code{\link{mdx_nativize_set}()},
\code{\link{mdx_non_empty}()},
\code{\link{mdx_non_empty_cross_join}()},
\code{\link{mdx_unorder}()}
}
\concept{set-operations}
