% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-helpers.R
\name{mdx_visual_totals}
\alias{mdx_visual_totals}
\title{MDX VisualTotals function}
\usage{
mdx_visual_totals(set, pattern = NULL)
}
\arguments{
\item{set}{Set expression.}

\item{pattern}{Optional pattern string for the visual totals caption.}
}
\value{
MDX VisualTotals expression string.
}
\description{
Returns a set generated by dynamically totaling child members
in a specified set.
}
