% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx-helpers.R
\name{mdx_yoy_growth}
\alias{mdx_yoy_growth}
\title{Build expression: year-over-year growth}
\usage{
mdx_yoy_growth(measure, time_hierarchy)
}
\arguments{
\item{measure}{Measure unique name.}

\item{time_hierarchy}{Time hierarchy unique name.}
}
\value{
MDX expression string for (current - prior) / prior.
}
\description{
Build expression: year-over-year growth
}
