% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-measure.R
\name{Measure}
\alias{Measure}
\title{Measure}
\description{
R6 class representing an XMLA measure.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog_name}}{Catalog name.}

\item{\code{cube_name}}{Cube name.}

\item{\code{measure_name}}{Measure name.}

\item{\code{measure_unique_name}}{Measure unique name.}

\item{\code{measure_caption}}{Measure caption.}

\item{\code{measure_aggregator}}{Aggregator type.}

\item{\code{data_type}}{Data type.}

\item{\code{description}}{Description.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Measure-new}{\code{Measure$new()}}
\item \href{#method-Measure-print}{\code{Measure$print()}}
\item \href{#method-Measure-clone}{\code{Measure$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Measure-new"></a>}}
\if{latex}{\out{\hypertarget{method-Measure-new}{}}}
\subsection{Method \code{new()}}{
Create a Measure.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$new(
  catalog_name = NA_character_,
  cube_name = NA_character_,
  measure_name = NA_character_,
  measure_unique_name = NA_character_,
  measure_caption = NA_character_,
  measure_aggregator = NA_character_,
  data_type = NA_character_,
  description = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalog_name}}{Catalog.}

\item{\code{cube_name}}{Cube.}

\item{\code{measure_name}}{Name.}

\item{\code{measure_unique_name}}{Unique name.}

\item{\code{measure_caption}}{Caption.}

\item{\code{measure_aggregator}}{Aggregator.}

\item{\code{data_type}}{Data type.}

\item{\code{description}}{Description.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Measure-print"></a>}}
\if{latex}{\out{\hypertarget{method-Measure-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Measure-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Measure-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Measure$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
