% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-measure.R
\name{Measure_from_node}
\alias{Measure_from_node}
\title{Create a Measure from an XML row node}
\usage{
Measure_from_node(node)
}
\arguments{
\item{node}{An xml2 node.}
}
\value{
A Measure object.
}
\description{
Create a Measure from an XML row node
}
\keyword{internal}
