% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-member.R
\name{Member}
\alias{Member}
\title{Member}
\description{
R6 class representing an XMLA member.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog_name}}{Catalog name.}

\item{\code{cube_name}}{Cube name.}

\item{\code{dimension_unique_name}}{Dimension unique name.}

\item{\code{hierarchy_unique_name}}{Hierarchy unique name.}

\item{\code{level_unique_name}}{Level unique name.}

\item{\code{member_name}}{Member name.}

\item{\code{member_unique_name}}{Member unique name.}

\item{\code{member_caption}}{Member caption.}

\item{\code{member_type}}{Member type.}

\item{\code{description}}{Description.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Member-new}{\code{Member$new()}}
\item \href{#method-Member-print}{\code{Member$print()}}
\item \href{#method-Member-clone}{\code{Member$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Member-new"></a>}}
\if{latex}{\out{\hypertarget{method-Member-new}{}}}
\subsection{Method \code{new()}}{
Create a Member.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Member$new(
  catalog_name = NA_character_,
  cube_name = NA_character_,
  dimension_unique_name = NA_character_,
  hierarchy_unique_name = NA_character_,
  level_unique_name = NA_character_,
  member_name = NA_character_,
  member_unique_name = NA_character_,
  member_caption = NA_character_,
  member_type = NA_character_,
  description = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalog_name}}{Catalog.}

\item{\code{cube_name}}{Cube.}

\item{\code{dimension_unique_name}}{Dimension unique name.}

\item{\code{hierarchy_unique_name}}{Hierarchy unique name.}

\item{\code{level_unique_name}}{Level unique name.}

\item{\code{member_name}}{Name.}

\item{\code{member_unique_name}}{Unique name.}

\item{\code{member_caption}}{Caption.}

\item{\code{member_type}}{Type.}

\item{\code{description}}{Description.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Member-print"></a>}}
\if{latex}{\out{\hypertarget{method-Member-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Member$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Member-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Member-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Member$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
