% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-member.R
\name{Member_from_node}
\alias{Member_from_node}
\title{Create a Member from an XML row node}
\usage{
Member_from_node(node)
}
\arguments{
\item{node}{An xml2 node.}
}
\value{
A Member object.
}
\description{
Create a Member from an XML row node
}
\keyword{internal}
