% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-helpers.R
\name{null_if_empty}
\alias{null_if_empty}
\title{Return NULL if value is empty or NA}
\usage{
null_if_empty(x)
}
\arguments{
\item{x}{A value to check.}
}
\value{
The value or NULL.
}
\description{
Return NULL if value is empty or NA
}
