% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-helpers.R
\name{optional_value}
\alias{optional_value}
\title{Return default if value is NULL or NA}
\usage{
optional_value(x, default = NULL)
}
\arguments{
\item{x}{A value to check.}

\item{default}{The default value.}
}
\value{
The value or the default.
}
\description{
Return default if value is NULL or NA
}
