% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-execute.R
\name{parse_axis}
\alias{parse_axis}
\title{Parse a single Axis node}
\usage{
parse_axis(axis_node)
}
\arguments{
\item{axis_node}{An xml2 node for an Axis element.}
}
\value{
An Axis object.
}
\description{
Parse a single Axis node
}
\keyword{internal}
