% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-execute.R
\name{parse_cell_data}
\alias{parse_cell_data}
\title{Parse CellData node}
\usage{
parse_cell_data(celldata_node)
}
\arguments{
\item{celldata_node}{An xml2 node for the CellData element.}
}
\value{
A CellData object.
}
\description{
Parse CellData node
}
\keyword{internal}
