% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-execute.R
\name{parse_multidimensional_response}
\alias{parse_multidimensional_response}
\title{Parse a multidimensional Execute response}
\usage{
parse_multidimensional_response(xml)
}
\arguments{
\item{xml}{An xml2 document.}
}
\value{
An ExecuteResult object.
}
\description{
Parse a multidimensional Execute response
}
\keyword{internal}
