% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-property.R
\name{Property}
\alias{Property}
\title{Property}
\description{
R6 class representing an XMLA property.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog_name}}{Catalog name.}

\item{\code{property_name}}{Property name.}

\item{\code{property_description}}{Property description.}

\item{\code{property_type}}{Property type.}

\item{\code{property_access_type}}{Access type.}

\item{\code{is_required}}{Whether property is required.}

\item{\code{value}}{Current value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Property-new}{\code{Property$new()}}
\item \href{#method-Property-print}{\code{Property$print()}}
\item \href{#method-Property-clone}{\code{Property$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Property-new"></a>}}
\if{latex}{\out{\hypertarget{method-Property-new}{}}}
\subsection{Method \code{new()}}{
Create a Property.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Property$new(
  catalog_name = NA_character_,
  property_name = NA_character_,
  property_description = NA_character_,
  property_type = NA_character_,
  property_access_type = NA_character_,
  is_required = NA,
  value = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalog_name}}{Catalog.}

\item{\code{property_name}}{Name.}

\item{\code{property_description}}{Description.}

\item{\code{property_type}}{Type.}

\item{\code{property_access_type}}{Access type.}

\item{\code{is_required}}{Required flag.}

\item{\code{value}}{Value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Property-print"></a>}}
\if{latex}{\out{\hypertarget{method-Property-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Property$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Property-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Property-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Property$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
