% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-property.R
\name{Property_from_node}
\alias{Property_from_node}
\title{Create a Property from an XML row node}
\usage{
Property_from_node(node)
}
\arguments{
\item{node}{An xml2 node.}
}
\value{
A Property object.
}
\description{
Create a Property from an XML row node
}
\keyword{internal}
