% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap-objects.R
\name{to_xml_string}
\alias{to_xml_string}
\title{Serialize a SOAP Envelope to an XML string}
\usage{
to_xml_string(envelope)
}
\arguments{
\item{envelope}{A node descriptor as returned by \code{soap_envelope}.}
}
\value{
A single XML string.
}
\description{
Serialize a SOAP Envelope to an XML string
}
