% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-utils.R
\name{xml_child_integer}
\alias{xml_child_integer}
\title{Extract integer value of a child element by local name}
\usage{
xml_child_integer(node, name)
}
\arguments{
\item{node}{An xml2 node.}

\item{name}{Local element name to find.}
}
\value{
Integer scalar or NA_integer_.
}
\description{
Extract integer value of a child element by local name
}
\keyword{internal}
