% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-utils.R
\name{xml_child_logical}
\alias{xml_child_logical}
\title{Extract logical value of a child element by local name}
\usage{
xml_child_logical(node, name)
}
\arguments{
\item{node}{An xml2 node.}

\item{name}{Local element name to find.}
}
\value{
Logical scalar or NA.
}
\description{
Extract logical value of a child element by local name
}
\keyword{internal}
