% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-utils.R
\name{xml_child_text}
\alias{xml_child_text}
\title{Extract text content of a child element by local name}
\usage{
xml_child_text(node, name)
}
\arguments{
\item{node}{An xml2 node.}

\item{name}{Local element name to find.}
}
\value{
Character scalar or NA_character_.
}
\description{
Extract text content of a child element by local name
}
\keyword{internal}
