% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-utils.R
\name{xml_get_discover_rows}
\alias{xml_get_discover_rows}
\title{Extract row nodes from a DiscoverResponse}
\usage{
xml_get_discover_rows(xml)
}
\arguments{
\item{xml}{An xml2 document from a Discover response.}
}
\value{
A nodeset of row elements.
}
\description{
Extract row nodes from a DiscoverResponse
}
\keyword{internal}
