% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-utils.R
\name{xml_node_to_list}
\alias{xml_node_to_list}
\title{Convert all children of a node to a named list}
\usage{
xml_node_to_list(node)
}
\arguments{
\item{node}{An xml2 node.}
}
\value{
A named list of character values.
}
\description{
Convert all children of a node to a named list
}
\keyword{internal}
