% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-utils.R
\name{xml_nodes_to_df}
\alias{xml_nodes_to_df}
\title{Convert a list of XML nodes to a data.frame}
\usage{
xml_nodes_to_df(nodes)
}
\arguments{
\item{nodes}{A list of xml2 nodes (each representing a row).}
}
\value{
A data.frame with one row per node.
}
\description{
Convert a list of XML nodes to a data.frame
}
\keyword{internal}
