% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap-objects.R
\name{xml_ns_detect}
\alias{xml_ns_detect}
\title{Detect XML namespace prefixes in an XMLA response}
\usage{
xml_ns_detect(doc)
}
\arguments{
\item{doc}{An xml2 document.}
}
\value{
A named character vector mapping canonical names to namespace URIs.
}
\description{
Maps document namespace URIs to canonical names used for XPath queries.
}
