% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-utils.R
\name{xmla_decode_names}
\alias{xmla_decode_names}
\title{Decode XMLA _xHHHH_ encoded names}
\usage{
xmla_decode_names(x)
}
\arguments{
\item{x}{Character vector to decode.}
}
\value{
Decoded character vector.
}
\description{
Decode XMLA _xHHHH_ encoded names
}
\keyword{internal}
