% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap-objects.R
\docType{data}
\name{XmlNamespaces}
\alias{XmlNamespaces}
\title{SOAP/XMLA Namespaces}
\format{
An object of class \code{list} of length 2.
}
\usage{
XmlNamespaces
}
\description{
Namespace registry used by all SOAP/XMLA objects.
}
\keyword{datasets}
